refactor-place-1: GetAllPlaces Payload Class 재설계
