+++ app/gourmet-place-api/src/main/kotlin/com/gourmet/service/place/api/serialize/PlaceLocation.kt
@@ -6,6 +6,11 @@ data class PlaceLocation(
     val latitude: Double,
     val longitude: Double
 ) {
+    fun toGeospatialPoint() = GeospatialPoint(
+        latitude,
+        longitude
+    )
+
     companion object {
         fun fromGeospatialPoint(geoPoint: GeospatialPoint) =
             PlaceLocation(geoPoint.getLatitude(), geoPoint.getLongitude())
