+++ core/gourmet-place-core/src/main/kotlin/com/gourmet/service/place/core/usecase/dto/GetAllPlacesOption.kt
@@ -1,7 +1,9 @@
 package com.gourmet.service.place.core.usecase.dto
 
+import com.gourmet.service.common.type.GeospatialPoint
 import com.gourmet.service.common.type.PagingInformation
 
 data class GetAllPlacesOption(
+    val location: GeospatialPoint? = null,
     val paging: PagingInformation? = null,
 )
